+++ src/modules/gatsby-source-url/createImgixGatsbyImageDataFieldConfig.ts
@@ -1,11 +1,10 @@
 import ImgixClient from '@imgix/js-core';
 import { stripIndent } from 'common-tags';
 import { Do } from 'fp-ts-contrib/lib/Do';
-import { pipe } from 'fp-ts/lib/function';
-import * as T from 'fp-ts/lib/Task';
-import * as TE from 'fp-ts/lib/TaskEither';
+import { pipe } from 'fp-ts/function';
+import * as T from 'fp-ts/Task';
+import * as TE from 'fp-ts/TaskEither';
 import { GatsbyCache } from 'gatsby';
-import { FluidObject } from 'gatsby-image';
 import {
   generateImageData,
   getLowResolutionImageURL,
@@ -14,30 +13,31 @@ import {
   ImageFormat,
 } from 'gatsby-plugin-image';
 import { getGatsbyImageFieldConfig } from 'gatsby-plugin-image/graphql-utils';
+import { GraphQLFieldResolver } from 'gatsby/graphql';
 import {
-  GraphQLFieldConfig,
-  GraphQLFieldResolver,
-  GraphQLFloat,
-  GraphQLInt,
-  GraphQLObjectType,
-} from 'gatsby/graphql';
+  ComposeInputTypeDefinition,
+  ObjectTypeComposerArgumentConfigMapDefinition,
+  ObjectTypeComposerFieldConfigAsObjectDefinition,
+} from 'graphql-compose';
+import { TypeAsString } from 'graphql-compose/lib/TypeMapper';
 import R from 'ramda';
 import {
   fetchImgixBase64Image,
   fetchImgixDominantColor,
 } from '../../api/fetchBase64Image';
+import { createExternalHelper } from '../../common/createExternalHelper';
 import { TaskOptionFromTE } from '../../common/fpTsUtils';
+import { IImgixURLBuilder } from '../../common/imgix-js-core-wrapper';
 import {
   ImgixSourceDataResolver,
   resolveUrlFromSourceData,
   taskEitherFromSourceDataResolver,
 } from '../../common/utils';
 import { IImgixParams, ImgixUrlParams } from '../../publicTypes';
-import { ImgixParamsInputType, ImgixPlaceholderType } from './graphqlTypes';
 import { resolveDimensions } from './resolveDimensions';
 
 const generateImageSource = (
-  client: ImgixClient,
+  client: IImgixURLBuilder,
 ): IGatsbyImageHelperArgs['generateImageSource'] => (
   imageName,
   width,
@@ -54,25 +54,66 @@ const generateImageSource = (
   return { width, height, format: 'auto', src };
 };
 
+export type IBuildGatsbyImageDataBaseArgs = {
+  resolverArgs: IImgixGatsbyImageDataArgsResolved;
+  url: string;
+  dimensions: {
+    width: number;
+    height: number;
+  };
+  defaultParams?: Partial<IImgixParams>;
+  imgixClient: IImgixURLBuilder;
+};
+
+export const buildGatsbyImageDataBaseArgs = ({
+  resolverArgs,
+  url,
+  dimensions: { width, height },
+  defaultParams,
+  imgixClient,
+}: IBuildGatsbyImageDataBaseArgs) =>
+  ({
+    ...resolverArgs,
+    pluginName: `@imgix/gatsby`,
+    filename: url,
+    sourceMetadata: { width, height, format: 'auto' as ImageFormat },
+    // TODO: use breakpoints helper from gatsby-plugin-image hook
+    breakpoints:
+      resolverArgs.breakpoints ??
+      ImgixClient.targetWidths(
+        resolverArgs.srcSetMinWidth,
+        resolverArgs.srcSetMaxWidth,
+        resolverArgs.widthTolerance,
+      ),
+    formats: ['auto'] as ImageFormat[],
+    generateImageSource: generateImageSource(imgixClient),
+    options: {
+      imgixParams: {
+        ...defaultParams,
+        ...resolverArgs.imgixParams,
+      },
+    },
+  } as const);
+
 const resolveGatsbyImageData = <TSource>({
   resolveUrl,
   imgixClient,
   resolveWidth = () => undefined,
   resolveHeight = () => undefined,
   cache,
+  defaultParams,
 }: {
-  imgixClient: ImgixClient;
+  imgixClient: IImgixURLBuilder;
   resolveUrl: ImgixSourceDataResolver<TSource, string>;
-  resolveWidth?: ImgixSourceDataResolver<TSource, number>;
-  resolveHeight?: ImgixSourceDataResolver<TSource, number>;
+  resolveWidth?: ImgixSourceDataResolver<TSource, number | undefined>;
+  resolveHeight?: ImgixSourceDataResolver<TSource, number | undefined>;
   cache: GatsbyCache;
   defaultParams?: Partial<IImgixParams>;
-  type?: GraphQLObjectType<FluidObject>;
 }): GraphQLFieldResolver<
   TSource,
   unknown,
   IImgixGatsbyImageDataArgsResolved
-> => async (rootValue, args): Promise<IGatsbyImageData | undefined> => {
+> => async (rootValue, resolverArgs): Promise<IGatsbyImageData | undefined> => {
   return pipe(
     Do(TE.taskEither)
       .sequenceSL(() => ({
@@ -97,39 +138,26 @@ const resolveGatsbyImageData = <TSource>({
           client: imgixClient,
         }),
       )
-      .letL(
-        'baseImageDataArgs',
-        ({ url, dimensions: { width, height } }) =>
-          ({
-            ...args,
-            pluginName: `@imgix/gatsby`,
-            filename: url,
-            sourceMetadata: { width, height, format: 'auto' as ImageFormat },
-            // TODO: use breakpoints helper from gatsby-plugin-image hook
-            breakpoints:
-              args.breakpoints ??
-              ImgixClient.targetWidths(
-                args.srcSetMinWidth,
-                args.srcSetMaxWidth,
-                args.widthTolerance,
-              ),
-            formats: ['auto'] as ImageFormat[],
-            generateImageSource: generateImageSource(imgixClient),
-            options: {
-              imgixParams: args.imgixParams,
-            },
-          } as const),
+      .letL('baseImageDataArgs', ({ url, dimensions }) =>
+        buildGatsbyImageDataBaseArgs({
+          url,
+          dimensions,
+          resolverArgs,
+          defaultParams,
+          imgixClient,
+        }),
       )
       .bindL('placeholderData', ({ url, baseImageDataArgs }) => {
-        if (args.placeholder === 'blurred') {
+        if (resolverArgs.placeholder === 'blurred') {
           return pipe(
             getLowResolutionImageURL({
               ...baseImageDataArgs,
               options: {
                 ...baseImageDataArgs,
                 imgixParams: {
-                  ...args.imgixParams,
-                  ...args.placeholderImgixParams,
+                  ...defaultParams,
+                  ...resolverArgs.imgixParams,
+                  ...resolverArgs.placeholderImgixParams,
                 },
               },
             }),
@@ -139,12 +167,13 @@ const resolveGatsbyImageData = <TSource>({
             })),
           );
         }
-        if (args.placeholder === 'dominantColor') {
+        if (resolverArgs.placeholder === 'dominantColor') {
           return pipe(
             fetchImgixDominantColor(cache)((params) =>
               imgixClient.buildURL(url, {
-                ...args.imgixParams,
-                ...args.placeholderImgixParams,
+                ...defaultParams,
+                ...resolverArgs.imgixParams,
+                ...resolverArgs.placeholderImgixParams,
                 ...params,
               }),
             ),
@@ -170,15 +199,20 @@ export const createImgixGatsbyImageFieldConfig = <TSource, TContext = {}>({
   imgixClient,
   resolveUrl,
   defaultParams,
+  paramsInputType,
+  placeholderEnumType,
+  resolveWidth,
+  resolveHeight,
 }: {
-  imgixClient: ImgixClient;
+  imgixClient: IImgixURLBuilder;
   resolveUrl: ImgixSourceDataResolver<TSource, string>;
-  resolveWidth?: ImgixSourceDataResolver<TSource, number>;
-  resolveHeight?: ImgixSourceDataResolver<TSource, number>;
+  resolveWidth?: ImgixSourceDataResolver<TSource, number | undefined>;
+  resolveHeight?: ImgixSourceDataResolver<TSource, number | undefined>;
   cache: GatsbyCache;
   defaultParams?: Partial<IImgixParams>;
-  type?: GraphQLObjectType<FluidObject>;
-}): GraphQLFieldConfig<
+  paramsInputType: ComposeInputTypeDefinition;
+  placeholderEnumType: TypeAsString;
+}): ObjectTypeComposerFieldConfigAsObjectDefinition<
   TSource,
   TContext,
   IImgixGatsbyImageDataArgsResolved
@@ -189,9 +223,17 @@ export const createImgixGatsbyImageFieldConfig = <TSource, TContext = {}>({
       imgixClient,
       resolveUrl,
       defaultParams,
+      resolveWidth,
+      resolveHeight,
     }) as GraphQLFieldResolver<TSource, TContext>, // TODO: remove cast when PR to Gatsby has been merged
     {},
-  );
+  ) as ObjectTypeComposerFieldConfigAsObjectDefinition<
+    TSource,
+    TContext,
+    IBuiltinGatsbyImageDataArgs
+  >;
+
+  const defaultArgs = defaultConfig.args as ObjectTypeComposerArgumentConfigMapDefinition<IBuiltinGatsbyImageDataArgs>;
 
   // ⚠️ KEEP THESE IN SYNC WITH IImgixGatsbyImageDataArgs!! ⚠️
   const modifiedConfig = {
@@ -208,24 +250,24 @@ export const createImgixGatsbyImageFieldConfig = <TSource, TContext = {}>({
           'sizes',
           'backgroundColor',
         ],
-        defaultConfig.args,
+        defaultArgs,
       ),
       imgixParams: {
-        type: ImgixParamsInputType,
+        type: paramsInputType,
         description: stripIndent`This argument is used to set parameters to instruct imgix to transform the image. 
         
         All of imgix's parameters can be found here: https://docs.imgix.com/apis/rendering
         `,
       },
       placeholderImgixParams: {
-        type: ImgixParamsInputType,
+        type: paramsInputType,
         description: stripIndent`This argument is used to set parameters to instruct imgix to transform the placeholder image. By default all parameters passed to 'imgixParams' are already set here, but this can be used to override or set extra parameters. 
         
         All of imgix's parameters can be found here: https://docs.imgix.com/apis/rendering
         `,
       },
       placeholder: {
-        type: ImgixPlaceholderType,
+        type: placeholderEnumType,
         description: stripIndent`
           Format of generated placeholder image, displayed while the main image loads.
           BLURRED: a blurred, low resolution image, encoded as a base64 data URI (default)
@@ -233,7 +275,7 @@ export const createImgixGatsbyImageFieldConfig = <TSource, TContext = {}>({
           NONE: no placeholder. Set "backgroundColor" to use a fixed background color.`,
       },
       widthTolerance: {
-        type: GraphQLFloat,
+        type: 'Float',
         description: stripIndent`
           This argument affects the breakpoints used for the srcsets, dictates the maximum tolerated size difference between an image's downloaded size and its rendered size. For example: setting this value to 0.1 means that an image will not render more than 10% larger or smaller than its native size. In practice, the image URLs generated for a width-based srcset attribute will grow by twice this rate. A lower tolerance means images will render closer to their native size (thereby increasing perceived image quality), but a large srcset list will be generated and consequently users may experience lower rates of cache-hit for pre-rendered images on your site.
 
@@ -241,14 +283,14 @@ export const createImgixGatsbyImageFieldConfig = <TSource, TContext = {}>({
         defaultValue: 0.08,
       },
       srcSetMinWidth: {
-        type: GraphQLInt,
+        type: 'Int',
         description: stripIndent`
           This argument determines the minimum srcset width that is generated. The default is 100px.
         `,
         defaultValue: 100,
       },
       srcSetMaxWidth: {
-        type: GraphQLInt,
+        type: 'Int',
         description: stripIndent`
           This argument determines the maximum srcset width that is generated, if the layout type is FULL_WIDTH. If the layout type is CONSTRAINED, the width argument will be used. Furthermore, in every case, the maximum srcset width is constrained by the width of the source image. The default is 8192px, which is the render limit of the imgix service.
         `,
@@ -260,15 +302,26 @@ export const createImgixGatsbyImageFieldConfig = <TSource, TContext = {}>({
   return modifiedConfig;
 };
 
-type IImgixGatsbyImageDataArgsResolved = {
-  layout: IGatsbyImageHelperArgs['layout'];
-  width: IGatsbyImageHelperArgs['width'];
-  height: IGatsbyImageHelperArgs['height'];
-  aspectRatio: IGatsbyImageHelperArgs['aspectRatio'];
+type IBuiltinGatsbyImageDataArgs = {
+  layout?: IGatsbyImageHelperArgs['layout'];
+  width?: IGatsbyImageHelperArgs['width'];
+  height?: IGatsbyImageHelperArgs['height'];
+  aspectRatio?: IGatsbyImageHelperArgs['aspectRatio'];
   // outputPixelDensities: IGatsbyImageHelperArgs['outputPixelDensities'];
-  breakpoints: IGatsbyImageHelperArgs['breakpoints'];
-  sizes: IGatsbyImageHelperArgs['sizes'];
-  backgroundColor: IGatsbyImageHelperArgs['backgroundColor'];
+  breakpoints?: IGatsbyImageHelperArgs['breakpoints'];
+  sizes?: IGatsbyImageHelperArgs['sizes'];
+  backgroundColor?: IGatsbyImageHelperArgs['backgroundColor'];
+};
+
+type IImgixGatsbyImageDataArgsResolved = {
+  layout?: IGatsbyImageHelperArgs['layout'];
+  width?: IGatsbyImageHelperArgs['width'];
+  height?: IGatsbyImageHelperArgs['height'];
+  aspectRatio?: IGatsbyImageHelperArgs['aspectRatio'];
+  // outputPixelDensities: IGatsbyImageHelperArgs['outputPixelDensities'];
+  breakpoints?: IGatsbyImageHelperArgs['breakpoints'];
+  sizes?: IGatsbyImageHelperArgs['sizes'];
+  backgroundColor?: IGatsbyImageHelperArgs['backgroundColor'];
   imgixParams?: ImgixUrlParams;
   placeholderImgixParams?: ImgixUrlParams;
   placeholder?: 'dominantColor' | 'blurred' | 'none';
@@ -276,3 +329,8 @@ type IImgixGatsbyImageDataArgsResolved = {
   srcSetMinWidth?: number;
   srcSetMaxWidth?: number;
 };
+
+export const createImgixGatsbyImageSchemaFieldConfig = createExternalHelper<
+  Parameters<typeof createImgixGatsbyImageFieldConfig>[0],
+  typeof createImgixGatsbyImageFieldConfig
+>(createImgixGatsbyImageFieldConfig);