+++ src/modules/gatsby-source-url/fieldNames.ts
@@ -1,0 +1,7 @@
+export const getTypeName = ({
+  namespace,
+  typeName,
+}: {
+  namespace?: string;
+  typeName: 'Fluid' | 'Fixed' | 'ParamsInput' | 'Placeholder';
+}) => `${namespace ?? 'Imgix'}${typeName}`;