+++ src/modules/gatsby-transform-node/gatsby-node.ts
@@ -1,28 +1,19 @@
 import { Do } from 'fp-ts-contrib/lib/Do';
-import * as E from 'fp-ts/lib/Either';
-import { pipe } from 'fp-ts/lib/pipeable';
+import * as E from 'fp-ts/Either';
+import { pipe } from 'fp-ts/pipeable';
 import { ICreateSchemaCustomizationHook, PatchedPluginOptions } from 'gatsby';
-import {
-  GraphQLFieldConfig,
-  GraphQLNonNull,
-  GraphQLObjectType,
-  GraphQLString,
-} from 'gatsby/graphql';
+import { GraphQLNonNull, GraphQLString } from 'gatsby/graphql';
 import { PathReporter } from 'io-ts/PathReporter';
 import * as R from 'ramda';
 import readPkgUp from 'read-pkg-up';
 import { IImgixGatsbyOptions, ImgixSourceType } from '../..';
-import { createImgixClient } from '../../common/imgix-js-core-wrapper';
+import {
+  createImgixURLBuilder,
+  IImgixURLBuilder,
+} from '../../common/imgix-js-core-wrapper';
 import { findPossibleURLsInNode } from '../../common/utils';
 import { ImgixGatsbyOptionsIOTS } from '../../publicTypes';
-import { createImgixFixedFieldConfig } from '../gatsby-source-url/createImgixFixedFieldConfig';
-import { createImgixFluidFieldConfig } from '../gatsby-source-url/createImgixFluidFieldConfig';
-import { createImgixGatsbyImageFieldConfig } from '../gatsby-source-url/createImgixGatsbyImageDataFieldConfig';
-import { createImgixUrlFieldConfig } from '../gatsby-source-url/createImgixUrlFieldConfig';
-import {
-  createImgixFixedType,
-  createImgixFluidType,
-} from '../gatsby-source-url/graphqlTypes';
+import { buildImgixGatsbyTypes } from '../gatsby-source-url/typeBuilder';
 
 function isStringArray(value: unknown): value is string[] {
   return (
@@ -102,29 +93,21 @@ const getPackageVersionE = () =>
     E.fromNullable(new Error('Could not read package version.')),
   );
 
-const setupImgixClientE = ({
+const setupImgixClient = ({
   options,
   packageVersion,
 }: {
   options: IImgixGatsbyOptions;
   packageVersion: string;
-}) =>
-  Do(E.either)
-    .bind(
-      'imgixClient',
-      createImgixClient({
-        domain: options.domain,
-        secureURLToken: options.secureURLToken,
-      }),
-    )
-    .doL(({ imgixClient }) => {
-      imgixClient.includeLibraryParam = false;
-      if (options.disableIxlibParam !== true) {
-        (imgixClient as any).settings.libraryParam = `gatsbySourceUrl-${packageVersion}`;
-      }
-      return E.right(imgixClient);
-    })
-    .return(R.prop('imgixClient'));
+}): IImgixURLBuilder =>
+  createImgixURLBuilder({
+    domain: options.domain,
+    secureURLToken: options.secureURLToken,
+    ixlib:
+      options.disableIxlibParam !== true
+        ? `gatsbySourceUrl-${packageVersion}`
+        : undefined,
+  });
 
 export const createSchemaCustomization: ICreateSchemaCustomizationHook<IImgixGatsbyOptions> = async (
   gatsbyContext,
@@ -135,61 +118,29 @@ export const createSchemaCustomization: ICreateSchemaCustomizationHook<IImgixGat
       .bind('options', decodeOptionsE(_options))
 
       .bind('packageVersion', getPackageVersionE())
-      .bindL('imgixClient', ({ options, packageVersion }) =>
-        setupImgixClientE({ options, packageVersion }),
-      )
-      .let(
-        'imgixFixedType',
-        createImgixFixedType({
-          name: 'ImgixFixed',
-          cache: gatsbyContext.cache,
-        }),
-      )
-      .let(
-        'imgixFluidType',
-        createImgixFluidType({
-          name: 'ImgixFluid',
-          cache: gatsbyContext.cache,
-        }),
+      .letL('imgixClient', ({ options, packageVersion }) =>
+        setupImgixClient({ options, packageVersion }),
       )
       .letL(
-        'imgixImageType',
-        ({
-          imgixFluidType,
-          imgixFixedType,
-          imgixClient,
-          options: { defaultImgixParams },
-        }) =>
-          new GraphQLObjectType<IRootSource, {}>({
-            name: 'ImgixImage',
-            fields: {
-              url: createImgixUrlFieldConfig({
-                resolveUrl: R.prop('rawURL'),
-                imgixClient,
-                defaultParams: defaultImgixParams,
-              }),
-              fluid: createImgixFluidFieldConfig({
-                type: imgixFluidType,
-                cache: gatsbyContext.cache,
-                imgixClient,
-                resolveUrl: R.prop('rawURL'),
-                defaultParams: defaultImgixParams,
-              }) as GraphQLFieldConfig<IRootSource, {}, Record<string, any>>,
-              fixed: createImgixFixedFieldConfig({
-                type: imgixFixedType,
-                cache: gatsbyContext.cache,
-                imgixClient,
-                resolveUrl: R.prop('rawURL'),
-                defaultParams: defaultImgixParams,
-              }) as GraphQLFieldConfig<IRootSource, {}, Record<string, any>>,
-              gatsbyImageData: createImgixGatsbyImageFieldConfig({
-                cache: gatsbyContext.cache,
-                imgixClient,
-                resolveUrl: R.prop('rawURL'),
-                defaultParams: defaultImgixParams,
-              }) as GraphQLFieldConfig<IRootSource, {}, Record<string, any>>,
-            },
+        'typesAndFields',
+        ({ imgixClient, options: { defaultImgixParams } }) =>
+          buildImgixGatsbyTypes<{ rawURL: string }>({
+            cache: gatsbyContext.cache,
+            imgixClient,
+            resolveUrl: R.prop('rawURL'),
+            defaultParams: defaultImgixParams,
           }),
+      )
+      .letL('imgixImageType', ({ typesAndFields }) =>
+        gatsbyContext.schema.buildObjectType({
+          name: 'ImgixImage',
+          fields: {
+            url: typesAndFields.fields.url,
+            fluid: typesAndFields.fields.fluid,
+            fixed: typesAndFields.fields.fixed,
+            gatsbyImageData: typesAndFields.fields.gatsbyImageData,
+          },
+        }),
       )
       .letL(
         'fieldTypes',
@@ -201,9 +152,9 @@ export const createSchemaCustomization: ICreateSchemaCustomizationHook<IImgixGat
                 [fieldOptions.fieldName]: {
                   type:
                     'getURLs' in fieldOptions
-                      ? `[${imgixImageType.name}]`
-                      : imgixImageType.name,
-                  resolve: (node: unknown) => {
+                      ? `[${imgixImageType.config.name}]`
+                      : imgixImageType.config.name,
+                  resolve: (node: unknown): { rawURL: string | string[] } => {
                     const rawURLE = getFieldValue({
                       fieldOptions,
                       node,
@@ -246,12 +197,12 @@ ${urlPathsFound
             }),
           ),
       )
-      .letL('rootType', ({ imgixImageType }) =>
+      .letL(`rootType`, ({ imgixImageType }) =>
         gatsbyContext.schema.buildObjectType({
           name: 'Query',
           fields: {
             imgixImage: {
-              type: imgixImageType,
+              type: imgixImageType.config.name,
               resolve(
                 _: any,
                 args: Record<string, unknown>,
@@ -272,24 +223,37 @@ ${urlPathsFound
           },
         }),
       )
+      // prettier-ignore
       .doL(
         ({
-          imgixFixedType,
-          imgixFluidType,
+          typesAndFields,
           imgixImageType,
           rootType,
           fieldTypes,
           options: {
-            // TODO: handle
-            // namespace,
-            // TODO: handle
-            // defaultPlaceholderImgixParams,
-          },
+          // TODO: handle
+          // namespace,
+          // TODO: handle
+          // prettier-ignore
+        // defaultPlaceholderImgixParams
+      },
         }) =>
           E.tryCatch(
             () => {
               const { createTypes } = gatsbyContext.actions;
-              createTypes([imgixFixedType, imgixFluidType]);
+              createTypes(
+                typesAndFields.types.map(gatsbyContext.schema.buildObjectType),
+              );
+              createTypes(
+                typesAndFields.enumTypes.map(
+                  gatsbyContext.schema.buildEnumType,
+                ),
+              );
+              createTypes(
+                typesAndFields.inputTypes.map(
+                  gatsbyContext.schema.buildInputObjectType,
+                ),
+              );
               createTypes(fieldTypes);
               createTypes(imgixImageType);
               createTypes(rootType);